introduce separation of axes in frame browser
user testing shows pan operations in both time and space are confusing.
to simplify, we isolate touch movement along an active axis.

activating an axis requires the pan gesture to translate beyond a
threshold. once active, discrete impulse events are emitted when pan
velocity exceeds an axis-specific threshold.

two sliders represent the current value of time and space. both are
currently rendered horizontally. this is not ideal for time, which is
interactive vertically.